Rename atom to recoil
